refactor: move volume rendering shaders to module
